*This commit might add a very small memory leak that needs confirmation * Optimize text rendering, to not create TTF textures all the time Only update the strings when something actually changes

+++ BlockFall/source/Game/GameField.h
@@ -61,7 +61,7 @@ struct GameField
     }
 
     GameRulesetsEnum ruleset;
-    int currentLevel = 0;
+    uint64_t currentLevel = 0;
     float currentSpeed = 0.0f;
     float previewWindowSize = 100.0f;
     float das; // Delay before auto-shift
